requirements.txt
qiskit-terra>=0.19.0
numpy>=1.17